combine staging workflows (#4188)
* combine staging workflows

* comment out build

* updates

* fix types

* fix prisma error

* add build for connect app

* separate build commands

* separate commands

* .

* cache connect modules

* fix cache key

* update staging to use name

* remove use of STAGE inside js code

* fix connecft api tests

* add postdeploy command

* fix deploy

* fix deployment url

* fix name

* use stage everywhere

* fix jq

* fix env var

* fix setting env var

* export raw output

* dont override

* deploy connect app based on label

* add quotes

* moreconditional

* fix conditional

* fix env that gets deployed

* fix keypair

* ..

* go back to image
