+++ .ebstalk.apps.env/connect.env
@@ -1,6 +1,4 @@
 # Format of this file is <env_var>=<value> key/value pair
 DATABASE_URL="{{pull:secretsmanager:/io.cv.app/prd/db:SecretString:database_url}}"
 
-ALLOWED_CONNECT_ORIGINS=
-
 # Frontend Secrets Start Here
