+++ .github/workflows/destroy_staging.yml
@@ -10,8 +10,8 @@ concurrency:
 jobs:
   clean-up:
     if: |
-      (github.event.action == 'unlabeled' && (github.event.label.name == ':rocket: deploy' || github.event.label.name == 'deploy-ceramic' || github.event.label.name == ':rocket: deploy-connect')) ||
-      (github.event.action == 'closed' && (contains(github.event.pull_request.labels.*.name, ':rocket: deploy') || contains(github.event.pull_request.labels.*.name, 'deploy-ceramic') || contains(github.event.pull_request.labels.*.name, ':rocket: deploy-connect')))
+      (github.event.action == 'unlabeled' && startsWith(github.event.label.name, ':rocket: deploy')) ||
+      (github.event.action == 'closed' && (contains(github.event.pull_request.labels.*.name, ':rocket: deploy') || contains(github.event.pull_request.labels.*.name, ':rocket: deploy-ceramic') || contains(github.event.pull_request.labels.*.name, ':rocket: deploy-connect')))
     runs-on: ubuntu-latest
     # wait until all deploys are complete to destroy the env
     concurrency: deploy-pr-app-${{ github.head_ref }}
@@ -40,7 +40,8 @@ jobs:
         run: |
           touch deploy.zip
           npm install aws-cdk-lib --no-audit --no-fund
-          npm run destroy:staging
+          npx cdk destroy \"stg-webapp-${STAGE}\" --force
+          npx cdk destroy \"stg-connect-${STAGE}\" --force
 
       - name: delete the github deployments and the corresponding environment
         uses: strumwolf/delete-deployment-environment@v1.1.0
