+++ apps/connect/app/page.tsx
@@ -2,6 +2,9 @@ import { redirect } from 'next/navigation';
 
 import { HomePage } from 'components/home/HomePage';
 
+// tell Next that this route loads dynamic data
+export const dynamic = 'force-dynamic';
+
 export default function Home() {
   const user = undefined; // server action
