+++ apps/connect/app/welcome/page.tsx
@@ -1,3 +1,3 @@
 export default async function Welcome() {
-  return null;
+  return <div>welcome world!</div>;
 }
