+++ package.json
@@ -77,14 +77,14 @@
     "build:storybook": "npx storybook build",
     "upgrade:storybook": "npx storybook@latest upgrade",
     "chromatic:ci": "chromatic --only-changed --exit-zero-on-changes",
-    "connect": "dotenv -e .env.local -- npm run start --workspace=apps/connect",
-    "connect:build": "dotenv -e .env.local -- npx next build ./apps/connect",
-    "connect:start:prod": "npm run start:prod --workspace=apps/connect",
-    "connect-api": "dotenv -e .env.local -- npm run start --workspace=apps/connect-api",
-    "connect-api:start:server": "dotenv -e .env.test.local -- npm run test:server --workspace=apps/connect-api",
-    "connect-api:test": "dotenv -e .env.test.local -- npm run test --workspace=apps/connect-api",
-    "connect-api:build": "dotenv -e .env.local -- npm run build --workspace=apps/connect-api",
-    "connect-api:start:prod": "npm run start:prod --workspace=apps/connect-api",
+    "connect": "dotenv -e .env.local -- npm run start --w=apps/connect",
+    "connect:build": "npm run build -w=apps/connect",
+    "connect:start:prod": "npm run start:prod --w=apps/connect",
+    "connect-api": "dotenv -e .env.local -- npm run start --w=apps/connect-api",
+    "connect-api:start:server": "dotenv -e .env.test.local -- npm run test:server --w=apps/connect-api",
+    "connect-api:test": "dotenv -e .env.test.local -- npm run test --w=apps/connect-api",
+    "connect-api:build": "dotenv -e .env.local -- npm run build --w=apps/connect-api",
+    "connect-api:start:prod": "npm run start:prod --w=apps/connect-api",
     "connect:all": "concurrently --names \"connect,connect-api\" -c \"green,yellow\" \"npm run connect\" \"npm run connect-api\"",
     "connect:all:build": "concurrently --names \"connect,connect-api\" -c \"green,yellow\" \"npm run connect:build\" \"npm run connect-api:build\""
   },
@@ -191,7 +191,6 @@
     "@synaps-io/verify-sdk": "^4.0.45",
     "@tanstack/react-query": "^5.28.9",
     "@typeform/api-client": "^2.3.1",
-    "@types/koa-router": "^7.4.8",
     "@uauth/js": "^2.4.0",
     "@wagmi/connectors": "^4.1.18",
     "@wagmi/core": "^2.6.9",
@@ -232,7 +231,7 @@
     "mailgun.js": "^10.2.1",
     "markdown-it": "^13.0.1",
     "marked": "^4.0.12",
-    "material-ui-popup-state": "^5.0.4",
+    "material-ui-popup-state": "^5.1.2",
     "millify": "^4.0.0",
     "mixpanel": "^0.17.0",
     "msw-storybook-addon": "^2.0.0",
@@ -351,6 +350,7 @@
     "@types/koa": "^2.15.0",
     "@types/koa__cors": "^5.0.0",
     "@types/koa__router": "^8.0.11",
+    "@types/koa-router": "^7.4.8",
     "@types/luxon": "^2.0.9",
     "@types/markdown-it": "^12.2.3",
     "@types/marked": "^4.0.2",
@@ -359,10 +359,10 @@
     "@types/papaparse": "^5.3.5",
     "@types/promise-retry": "^1.1.3",
     "@types/qrcode": "^1.5.5",
-    "@types/react": "^18.0.26",
+    "@types/react": "^18.3.3",
     "@types/react-copy-to-clipboard": "^5.0.4",
     "@types/react-csv": "^1.1.2",
-    "@types/react-dom": "^18.0.10",
+    "@types/react-dom": "^18.3.0",
     "@types/react-pdf": "^7.0.0",
     "@types/react-resizable": "^1.7.4",
     "@types/react-window": "^1.8.7",
@@ -423,7 +423,6 @@
     "webpack": "^5.76.0"
   },
   "overrides": {
-    "@types/react": "^18.0.26",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "typescript": "^5.4.2",
