+++ scripts/optimism/exportRetroCSVs.ts
@@ -8,8 +8,8 @@ import { uniq } from 'lodash';
 type SummaryRow = {};
 
 const summaryFile = './op-review-summary.csv';
-const fullReviewsummaryFile = './op-full-review-june-18.csv';
-const reviewersFile = './op-reviewers-june-18.csv';
+const fullReviewsummaryFile = './op-full-review-june-19.csv';
+const reviewersFile = './op-reviewers-june-19.csv';
 
 async function exportSummary() {
   const proposals = await prisma.proposal.findMany({
@@ -243,6 +243,6 @@ async function exportReviewers() {
   writeFileSync(reviewersFile, csvString);
 }
 
-// exportFullReviewSummary().catch(console.error);
+exportFullReviewSummary().catch(console.error);
 
 exportReviewers().catch(console.error);
